Consolidate reflow operation as new method.
